full_site/requirements.txt
fastapi
uvicorn
jinja2
requests
python-telegram-bot 
